# src/swaggerOptions.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "BLOG123",
    "description": "Express REST API for Blog"
  },
  "servers": [
    {
      "url": "/",
      "description": "Local Dev, or from Heroku"
    },
    {
      "url": "/api/",
      "description": "With docker-compose and nginx proxy"
    }
  ],
...
  "paths": {
    "/api/posts": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Get all posts",
        "security": [
          {"JwtAuth": []}
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/PostsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Posts"
        ],
        "security": [
          {"JwtAuth": []}
        ],
        "summary": "Create a new Post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/PostsBody"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/PostResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "PostResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "description": "the id of the post"
        },
        "authorId": {
          "type": "number",
          "description": "the id of the author of the post"
        },
        "title": {
          "type": "string",
          "description": "the title of the post"
        },
        "text": {
          "type": "string",
          "description": "the text of the post"
        },
        "imageUrl": {
          "type": "string",
          "description": "url to the image resource from the post"
        },
        "videoUrl": {
          "type": "string",
          "description": "url to the video resource from the post"
        }
      }
    },
    "PostBody": {
      "type": "object",
      "properties": {
        "authorId": {
          "type": "number",
          "required": true,
          "description": "the id of the author of the post"
        },
        "title": {
          "type": "string",
          "required": true,
          "description": "the title of the post"
        },
        "text": {
          "type": "string",
          "required": true,
          "description": "the text of the post"
        },
        "imageUrl": {
          "type": "string",
          "description": "url to the image resource from the post"
        },
        "videoUrl": {
          "type": "string",
          "description": "url to the video resource from the post"
        }
      }
    },
...
    "PostsBody": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/PostBody"
      }
    },
    "PostsResponse": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/PostResponse"
      }
    },
...
  }
}